Add basic side bar setup
