--- packages/ui/src/layouts/email-portal/dashborad/index.ts
@@ -1,3 +1,0 @@
-export * from './app-bar'
-export * from './side-bar'
-export * from './footer'
