+++ packages/ui/src/lib/providers/index.ts
@@ -1,0 +1,1 @@
+export * from './side-bar-provider'
