+++ packages/ui/src/styles/globals.css
@@ -4,77 +4,3 @@
 @source "../**/*.{ts,tsx}";
 @import "../../../configs/design-system/design-tokens.css";
 @import "tw-animate-css";
-
-@custom-variant dark (&:is(.dark *));
-
-/* Your existing CSS custom properties */
-:root {
-    --primary: #0798f3;
-    --primary-inverse: #9524d6;
-    /* ... all your other custom properties */
-    --sidebar: hsl(0 0% 98%);
-    --sidebar-foreground: hsl(240 5.3% 26.1%);
-    --sidebar-primary: hsl(240 5.9% 10%);
-    --sidebar-primary-foreground: hsl(0 0% 98%);
-    --sidebar-accent: hsl(240 4.8% 95.9%);
-    --sidebar-accent-foreground: hsl(240 5.9% 10%);
-    --sidebar-border: hsl(220 13% 91%);
-    --sidebar-ring: hsl(217.2 91.2% 59.8%);
-}
-
-/* Define Tailwind theme using CSS */
-@theme {
-    /* Map your semantic tokens to Tailwind utilities */
-    --color-primary: var(--primary);
-    --color-primary-inverse: var(--primary-inverse);
-
-    /* Surface colors that change with theme */
-    --color-surface-primary: var(--surface-primary);
-    --color-surface-secondary: var(--surface-secondary);
-
-    /* Content colors that change with theme */
-    --color-content-primary: var(--content-primary);
-    --color-content-secondary: var(--content-secondary);
-
-    /* Button-specific colors for your component */
-    --color-primary-foreground: var(--content-inverse-primary);
-    --color-destructive: var(--error);
-    --color-secondary: var(--bg-fill-secondary);
-    --color-secondary-foreground: var(--content-primary);
-    --color-background: var(--surface-primary);
-    --color-accent: var(--bg-fill-secondary);
-    --color-accent-foreground: var(--content-primary);
-    --color-input: var(--neutral-30);
-    --color-ring: var(--primary);
-}
-
-.dark {
-    --sidebar: hsl(240 5.9% 10%);
-    --sidebar-foreground: hsl(240 4.8% 95.9%);
-    --sidebar-primary: hsl(224.3 76.3% 48%);
-    --sidebar-primary-foreground: hsl(0 0% 100%);
-    --sidebar-accent: hsl(240 3.7% 15.9%);
-    --sidebar-accent-foreground: hsl(240 4.8% 95.9%);
-    --sidebar-border: hsl(240 3.7% 15.9%);
-    --sidebar-ring: hsl(217.2 91.2% 59.8%);
-}
-
-@theme inline {
-    --color-sidebar: var(--sidebar);
-    --color-sidebar-foreground: var(--sidebar-foreground);
-    --color-sidebar-primary: var(--sidebar-primary);
-    --color-sidebar-primary-foreground: var(--sidebar-primary-foreground);
-    --color-sidebar-accent: var(--sidebar-accent);
-    --color-sidebar-accent-foreground: var(--sidebar-accent-foreground);
-    --color-sidebar-border: var(--sidebar-border);
-    --color-sidebar-ring: var(--sidebar-ring);
-}
-
-@layer base {
-  * {
-    @apply border-border outline-ring/50;
-    }
-  body {
-    @apply bg-background text-foreground;
-    }
-}
